python/pyproject.toml
[tool.poetry]
name = "fabric"
version = "0.1.0"
description = "A Python library for the fabric system"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "fabric"}]

[tool.poetry.dependencies]
python = "^3.10"
eclipse-zenoh = "^0.11.0"
backoff = "^2.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.21.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"